requirements.txt
numpy==1.22.4
pandas==1.4.2
python-dotenv==0.20.0
requests==2.28.0
